Extract TreeDragDropHandler to UI/Helpers
Encapsulate all drag-drop state, event handlers, and visual helpers
into a standalone TreeDragDropHandler class with ITreeDragDropHost
interface for callbacks. Removes ~250 lines from code-behind.
